Fixes PUT operations with binary body
If you are uploading a file with a binary body, the existing hash256(request.body) code didn’t actually create the correct hash256. The bodyHash always is returned as the empty hash256.

However, with an AWS request, the body hash needs to be provided as a header called “x-aws-content-sha256”. Therefore, instead of trying to calculate this hash, simply use the one that is already present.